+++ stats.go
@@ -671,6 +671,38 @@ func Covariance(data1, data2 Float64Data) (float64, error) {
 	return ss * float64(l1) / float64(l1-1), nil
 }
 
+func Correlation(data1, data2 Float64Data) (float64, error) {
+
+	l1 := data1.Len()
+	l2 := data2.Len()
+
+	if l1 == 0 || l2 == 0 {
+		return 0, errors.New("Input data must not be empty")
+	}
+
+	if l1 != l2 {
+		return 0, errors.New("Input data must be same length")
+	}
+
+	var sum_xsq, sum_ysq, sum_cross float64
+
+	mean_x := data1.Get(0)
+	mean_y := data2.Get(0)
+
+	for i := 1; i < l1; i++ {
+		ratio := float64(i) / float64(i+1)
+		delta_x := data1.Get(i) - mean_x
+		delta_y := data2.Get(i) - mean_y
+		sum_xsq += delta_x * delta_x * ratio
+		sum_ysq += delta_y * delta_y * ratio
+		sum_cross += delta_x * delta_y * ratio
+		mean_x += delta_x / float64(i+1)
+		mean_y += delta_y / float64(i+1)
+	}
+
+	return sum_cross / (math.Sqrt(sum_xsq) * math.Sqrt(sum_ysq)), nil
+}
+
 // float64ToInt rounds a float64 to an int
 func float64ToInt(input float64) (output int) {
 	r, _ := Round(input, 0)